cpu: aarch64: consolidate jit_*_core_bf16cvt.hpp into jit_sve_512_core_bf16cvt.hpp
