+++ src/cpu/aarch64/jit_avx512_common_resampling.cpp
@@ -22,8 +22,8 @@
 #include "common/type_helpers.hpp"
 #include "common/utils.hpp"
 
-#include "cpu/aarch64/jit_avx512_core_bf16cvt.hpp"
 #include "cpu/aarch64/jit_generator.hpp"
+#include "cpu/aarch64/jit_sve_512_core_bf16cvt.hpp"
 
 #include "cpu/aarch64/jit_avx512_common_resampling.hpp"
