+++ src/cpu/aarch64/jit_uni_softmax.hpp
@@ -25,7 +25,7 @@
 #include "common/utils.hpp"
 
 #include "cpu/aarch64/cpu_isa_traits.hpp"
-#include "cpu/aarch64/jit_avx512_core_bf16cvt.hpp"
+#include "cpu/aarch64/jit_sve_512_core_bf16cvt.hpp"
 #include "cpu/cpu_softmax_pd.hpp"
 
 namespace dnnl {
